pyproject.toml
[tool.poetry]
name = "la_emissions_data"
version = "0.1.0"
description = "Transformed and enhanced version of the BEIS LA emissions dataset"
authors = []

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
data_common = { path = "src/data_common/", develop = true }
geopandas = "^0.11.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"

[tool.poetry.scripts]
project = 'la_emissions_data.__main__:main'
notebook = "data_common.notebookcli.__main__:run"
dataset = "data_common.dataset.__main__:run"

[tool.pyright]
include = ["src", "notebooks"]
exclude=["src/data_common/typings"]
stubPath="src/data_common/typings"
typeCheckingMode="basic"
reportPrivateImportUsage="warning"

[notebook.settings]
default_page_title = "Settings file defined title"

[tool.dataset]
dataset_dir = "data/packages"
publish_dir = "docs/"
publish_url = "https://mysociety.github.io/la_emissions_data/"
credit_text = "If you find this data useful, please let us know to help us make the case for future funding."
credit_url = "https://survey.alchemer.com/s3/6876792/Data-usage"